mail_service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
cachetools==5.3.2
    # via google-auth
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
google-api-core[grpc]==2.12.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-cloud-trace
google-api-python-client==2.107.0
    # via google-cloud-profiler
google-auth==2.23.4
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-profiler
google-auth-httplib2==0.1.1
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-cloud-profiler==4.1.0
    # via -r requirements.in
google-cloud-trace==1.11.3
    # via -r requirements.in
googleapis-common-protos==1.61.0
    # via
    #   google-api-core
    #   grpcio-status
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio==1.59.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpcio-health-checking
    #   grpcio-status
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-health-checking==1.59.2
    # via -r requirements.in
grpcio-status==1.59.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via opentelemetry-api
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
opentelemetry-api==1.20.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-distro==0.41b0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.20.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.20.0
    # via -r requirements.in
opentelemetry-instrumentation==0.41b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-grpc==0.41b0
    # via -r requirements.in
opentelemetry-proto==1.20.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.20.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.41b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
proto-plus==1.22.3
    # via google-cloud-trace
protobuf==4.25.0
    # via
    #   google-api-core
    #   google-cloud-profiler
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   grpcio-status
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.1.1
    # via httplib2
python-json-logger==2.0.7
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
rsa==4.9
    # via google-auth
typing-extensions==4.8.0
    # via opentelemetry-sdk
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.0.7
    # via requests
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
zipp==3.17.0
    # via importlib-metadata

